# api/docs/swagger.yaml
basePath: /api
definitions:
  api.JSONResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  api.ScanResultResponse:
    properties:
      error:
        type: string
      message:
        type: string
      result:
        type: object
      success:
        type: boolean
    type: object
  api.getAllNumbersResponse:
    properties:
      error:
        type: string
      message:
        type: string
      numbers:
        items:
          $ref: '#/definitions/number.Number'
        type: array
      success:
        type: boolean
    type: object
  api.healthResponse:
    properties:
      commit:
        type: string
      success:
        type: boolean
      version:
        type: string
    type: object
  number.Number:
    properties:
      carrier:
        type: string
      country:
        type: string
      countryCode:
        type: integer
      e164:
        type: string
      international:
        type: string
      local:
        type: string
      rawLocal:
        type: string
    type: object
  remote.GoogleSearchDork:
    properties:
      dork:
        type: string
      number:
        type: string
      url:
        type: string
    type: object
  remote.GoogleSearchResponse:
    properties:
      disposable_providers:
        items:
          $ref: '#/definitions/remote.GoogleSearchDork'
        type: array
      general:
        items:
          $ref: '#/definitions/remote.GoogleSearchDork'
        type: array
      individuals:
        items:
          $ref: '#/definitions/remote.GoogleSearchDork'
        type: array
      reputation:
        items:
          $ref: '#/definitions/remote.GoogleSearchDork'
        type: array
      social_media:
        items:
          $ref: '#/definitions/remote.GoogleSearchDork'
        type: array
    type: object
  remote.NumverifyScannerResponse:
    properties:
      carrier:
        type: string
      country_code:
        type: string
      country_name:
        type: string
      country_prefix:
        type: string
      international_format:
        type: string
      line_type:
        type: string
      local_format:
        type: string
      location:
        type: string
      number:
        type: string
      valid:
        type: boolean
    type: object
  remote.OVHScannerResponse:
    properties:
      city:
        type: string
      found:
        type: boolean
      number_range:
        type: string
      zip_code:
        type: string
    type: object
  remote.TempophoneScannerResponse:
    properties:
      Id:
        type: integer
      event_url:
        type: string
      found:
        type: boolean
    type: object
host: demo.phoneinfoga.crvx.fr
info:
  contact: {}
  description: Advanced information gathering & OSINT framework for phone numbers.
  license:
    name: GNU General Public License v3.0
    url: https://github.com/sundowndev/phoneinfoga/blob/master/LICENSE
  title: PhoneInfoga REST API
  version: v2
paths:
  /:
    get:
      operationId: healthCheck
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.healthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.JSONResponse'
      summary: Check if service is healthy.
      tags:
      - General
  /numbers:
    get:
      deprecated: true
      description: This route is actually not used yet.
      operationId: getAllNumbers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.getAllNumbersResponse'
      summary: Fetch all previously scanned numbers.
      tags:
      - Numbers
  /numbers/{number}/scan/googlesearch:
    get:
      operationId: googleSearchScan
      parameters:
      - description: Input phone number
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.ScanResultResponse'
            - properties:
                result:
                  $ref: '#/definitions/remote.GoogleSearchResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.JSONResponse'
      summary: Perform a scan using Google Search engine.
      tags:
      - Numbers
  /numbers/{number}/scan/local:
    get:
      operationId: localScan
      parameters:
      - description: Input phone number
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.ScanResultResponse'
            - properties:
                result:
                  $ref: '#/definitions/number.Number'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.JSONResponse'
      summary: Perform a scan using local phone number library.
      tags:
      - Numbers
  /numbers/{number}/scan/numverify:
    get:
      operationId: numverifyScan
      parameters:
      - description: Input phone number
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.ScanResultResponse'
            - properties:
                result:
                  $ref: '#/definitions/remote.NumverifyScannerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.JSONResponse'
      summary: Perform a scan using Numverify's API.
      tags:
      - Numbers
  /numbers/{number}/scan/ovh:
    get:
      operationId: ovhScan
      parameters:
      - description: Input phone number
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.ScanResultResponse'
            - properties:
                result:
                  $ref: '#/definitions/remote.OVHScannerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.JSONResponse'
      summary: Perform a scan using OVH's API.
      tags:
      - Numbers
  /numbers/{number}/scan/tempophone:
    get:
      operationId: tempophoneScan
      parameters:
      - description: Input phone number
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.ScanResultResponse'
            - properties:
                result:
                  $ref: '#/definitions/remote.TempophoneScannerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.JSONResponse'
      summary: Perform a scan using Tempophone's API.
      tags:
      - Numbers
  /numbers/{number}/validate:
    get:
      operationId: validate
      parameters:
      - description: Input phone number
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.JSONResponse'
      summary: Check if a number is valid and possible.
      tags:
      - Numbers
schemes:
- http
- https
swagger: "2.0"